fix: socket notifications
