+++ src/hooks/useTransactions.jsx
@@ -2,8 +2,21 @@ import {useMutation, useQueryClient} from '@tanstack/react-query'
 
 import {postRequest} from '../services/httpRequest'
 import notification from '../utils/notification'
+import socketIO from '../services/socket'
 
-const addTransaction = (body) => postRequest('/transactions/', body)
+const socket = socketIO()
+
+const addTransaction = async (values) => {
+  const response = await postRequest('/transactions/', values)
+
+  socket.emit('join_channel', values.user)
+  socket.emit('send_transaction', {
+    message: 'Transferencia recibida',
+    channel: values.user,
+  })
+
+  return response
+}
 
 // MUTATIONS
 const useCreateTransaction = () => {
